// src/app/pages/operaciones/training/register-interface.ts
export interface RegisterInterface {
  idCollaborator: number;

  namesCollaborator: string;
  lastnameCollaborator: string;

  idLeader: number;
  namesLeader: string;

  idRegion: number;
  descriptionRegion: string;

  idCommunity: number;
  descriptionCommunity: string;

  idFunctionalLeader: number;
  namesFunctionalLeader: string;

  dateAdmission: string;
}